requirements.txt
opencv-python
transformers
Pillow
tqdm
scikit-learn
matplotlib
streamlit
yt-dlp